Break out request URI into component parts.
Rather than divine component parts from user-provided string and incur
complexity overhead of input validation, this commit breaks out
component parts of the URI into fields of the request class so that
users can manually input this information to the right places.
This leaves room for an advanced client, wrapper, or factory method to
set fields automatically for users.