Add basic support for variable length responses.
This commit introduces support for variable length response parsing
based on the Content-Length header. If it is present, continue to
poll until the expected content length is satisfied. This is one
heurisitc for determining that a response is complete. This needs to
be refined to support chunked transfer encoding (which is often
mutually exclusive to content-length).